Interface/requirements.txt
google-api-python-client
kafka-python
    